+++ lib/core/internals/debugger.ts
@@ -17,12 +17,11 @@ export default class Debugger {
 
 	static _registerEvent( ...args: coreEventArgs<coreEventName> ) {
 		const eventName = args[0];
-		const templateName = args[2]['0'].template_name;
-		const contextData = args[2]['0'].ctx;
+		const meta = args[1];
 		if( args[1].config.debug ) {
 			log( 'HCL_EVENT: ', eventName );
-			log( 'HCL_TEMPLATE: ', templateName );
-			log( 'HCL_CTX: ', contextData );
+			log( '\n' );
+			log( 'IEVENT_DATA:', meta );
 		}
 	}